Restructured database models into separate app models
