+++ data_storage/urls.py
@@ -12,7 +12,7 @@ urlpatterns = [
     path('field-management/trials', views.list_trials, name='list_trials'),
     path('field-management/plots', views.list_plots, name='list_plots'),
     path('field-management/locations', views.list_locations, name='list_locations'),
-    path('field-management/locations', views.list_states_locations, name='locations_list'),
+    # path('field-management/locations', views.list_states_locations, name='locations_list'),
     path('observations/data', views.view_data, name='view_data'),
     path('observations/analyze', views.analyze_data, name='analyze_data'),
     path('observations/graph', views.graph_data, name='graph_data'),
