+++ ontology/templates/ontology/sop.html
@@ -28,7 +28,7 @@
         </tr>
         <tr>
             <td>
-                <a class="link-light" href="" download>
+                <a class="link-light" href="https://zenodo.org/records/10520249/files/TLI_PLP_10041247_sainfoin_cloning.pdf?download=1" download>
                     Maize plant height
                 </a>
             </td>
@@ -37,7 +37,7 @@
         </tr>
         <tr>
             <td>
-                <a class="link-light" href="" download>
+                <a class="link-light" href="https://zenodo.org/records/10009905/files/TLI_PLP_8412413_sop_creation_policies.pdf?download=1" download>
                     PGC plot image acquisition
                 </a>
             </td>
@@ -46,7 +46,7 @@
         </tr>
         <tr>
             <td>
-                <a class="link-light" href="" download>
+                <a class="link-light" href="https://zenodo.org/records/10520249/files/TLI_PLP_10041247_sainfoin_cloning.pdf?download=1" download>
                     LI-COR gas collar installation
                 </a>
             </td>
@@ -55,7 +55,7 @@
         </tr>
         <tr>
             <td>
-                <a class="link-light" href="" download>
+                <a class="link-light" href="https://zenodo.org/records/10009905/files/TLI_PLP_8412413_sop_creation_policies.pdf?download=1" download>
                     PGC aboveground biomass measurement
                 </a>
             </td>
